refactor: simplify requests/client, drop raw for now, fix missing schema context in request handlers (#429)
* cleanup multi request

* remove unused `raw` in client.
it was used in the past in Server Side Rendering to prepare data on the server before loading the client page.
however as it is just the same as non raw + S.Encode, it is not really valuable anyway.
In the futue we could reintroduce when needed, but instead as a `mode: "raw"` or `raw: true` option.

* fix: missing Schema Contexts in R